Add os.path.pure
