+++ Lib/posixpath/__init__.py
@@ -5,9 +5,6 @@ this module as os.path.  The "os.path" name is an alias for this
 module on Posix systems; on other systems (e.g. Windows),
 os.path provides the same operations in a manner specific to that
 platform, and is an alias to another module (e.g. ntpath).
-
-Some of this can actually be useful on non-Posix systems too, e.g.
-for manipulation of the pathname component of URLs.
 """
 
 # Strings representing various path-related bits and pieces.
@@ -25,162 +22,18 @@ devnull = '/dev/null'
 import os
 import sys
 import stat
-import genericpath
 from genericpath import *
+import posixpath.pure
+from posixpath.pure import *
 
-__all__ = ["normcase","isabs","join","splitdrive","splitroot","split","splitext",
-           "basename","dirname","commonprefix","getsize","getmtime",
-           "getatime","getctime","islink","exists","lexists","isdir","isfile",
-           "ismount", "expanduser","expandvars","normpath","abspath",
-           "samefile","sameopenfile","samestat",
-           "curdir","pardir","sep","pathsep","defpath","altsep","extsep",
-           "devnull","realpath","supports_unicode_filenames","relpath",
-           "commonpath", "isjunction","isdevdrive"]
-
-
-def _get_sep(path):
-    if isinstance(path, bytes):
-        return b'/'
-    else:
-        return '/'
-
-# Normalize the case of a pathname.  Trivial in Posix, string.lower on Mac.
-# On MS-DOS this may also turn slashes into backslashes; however, other
-# normalizations (such as optimizing '../' away) are not allowed
-# (another function should be defined to do that).
-
-def normcase(s):
-    """Normalize case of pathname.  Has no effect under Posix"""
-    return os.fspath(s)
-
-
-# Return whether a path is absolute.
-# Trivial in Posix, harder on the Mac or MS-DOS.
-
-def isabs(s):
-    """Test whether a path is absolute"""
-    s = os.fspath(s)
-    sep = _get_sep(s)
-    return s.startswith(sep)
-
-
-# Join pathnames.
-# Ignore the previous parts if a part is absolute.
-# Insert a '/' unless the first part is empty or already ends in '/'.
-
-def join(a, *p):
-    """Join two or more pathname components, inserting '/' as needed.
-    If any component is an absolute path, all previous path components
-    will be discarded.  An empty last part will result in a path that
-    ends with a separator."""
-    a = os.fspath(a)
-    sep = _get_sep(a)
-    path = a
-    try:
-        for b in p:
-            b = os.fspath(b)
-            if b.startswith(sep) or not path:
-                path = b
-            elif path.endswith(sep):
-                path += b
-            else:
-                path += sep + b
-    except (TypeError, AttributeError, BytesWarning):
-        genericpath._check_arg_types('join', a, *p)
-        raise
-    return path
-
-
-# Split a path in head (everything up to the last '/') and tail (the
-# rest).  If the path ends in '/', tail will be empty.  If there is no
-# '/' in the path, head  will be empty.
-# Trailing '/'es are stripped from head unless it is the root.
-
-def split(p):
-    """Split a pathname.  Returns tuple "(head, tail)" where "tail" is
-    everything after the final slash.  Either part may be empty."""
-    p = os.fspath(p)
-    sep = _get_sep(p)
-    i = p.rfind(sep) + 1
-    head, tail = p[:i], p[i:]
-    if head and head != sep*len(head):
-        head = head.rstrip(sep)
-    return head, tail
-
-
-# Split a path in root and extension.
-# The extension is everything starting at the last dot in the last
-# pathname component; the root is everything before that.
-# It is always true that root + ext == p.
-
-def splitext(p):
-    p = os.fspath(p)
-    if isinstance(p, bytes):
-        sep = b'/'
-        extsep = b'.'
-    else:
-        sep = '/'
-        extsep = '.'
-    return genericpath._splitext(p, sep, None, extsep)
-splitext.__doc__ = genericpath._splitext.__doc__
-
-# Split a pathname into a drive specification and the rest of the
-# path.  Useful on DOS/Windows/NT; on Unix, the drive is always empty.
-
-def splitdrive(p):
-    """Split a pathname into drive and path. On Posix, drive is always
-    empty."""
-    p = os.fspath(p)
-    return p[:0], p
-
-
-try:
-    from posix import _path_splitroot_ex as splitroot
-except ImportError:
-    def splitroot(p):
-        """Split a pathname into drive, root and tail.
-
-        The tail contains anything after the root."""
-        p = os.fspath(p)
-        if isinstance(p, bytes):
-            sep = b'/'
-            empty = b''
-        else:
-            sep = '/'
-            empty = ''
-        if p[:1] != sep:
-            # Relative path, e.g.: 'foo'
-            return empty, empty, p
-        elif p[1:2] != sep or p[2:3] == sep:
-            # Absolute path, e.g.: '/foo', '///foo', '////foo', etc.
-            return empty, sep, p[1:]
-        else:
-            # Precisely two leading slashes, e.g.: '//foo'. Implementation defined per POSIX, see
-            # https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap04.html#tag_04_13
-            return empty, p[:2], p[2:]
-
-
-# Return the tail (basename) part of a path, same as split(path)[1].
-
-def basename(p):
-    """Returns the final component of a pathname"""
-    p = os.fspath(p)
-    sep = _get_sep(p)
-    i = p.rfind(sep) + 1
-    return p[i:]
-
-
-# Return the head (dirname) part of a path, same as split(path)[0].
-
-def dirname(p):
-    """Returns the directory component of a pathname"""
-    p = os.fspath(p)
-    sep = _get_sep(p)
-    i = p.rfind(sep) + 1
-    head = p[:i]
-    if head and head != sep*len(head):
-        head = head.rstrip(sep)
-    return head
+__all__ = ["abspath", "altsep", "basename", "commonpath", "commonprefix",
+           "curdir", "defpath", "devnull", "dirname", "exists", "expanduser",
+           "expandvars", "extsep", "getatime", "getctime", "getmtime",
+           "getsize", "isabs", "isdevdrive", "isdir", "isfile", "isjunction",
+           "islink", "ismount", "join", "lexists", "normcase", "normpath",
+           "pardir", "pathsep", "realpath", "relpath", "samefile",
+           "sameopenfile", "samestat", "sep", "split", "splitdrive",
+           "splitext", "splitroot", "supports_unicode_filenames"]
 
 
 # Is a path a mount point?
@@ -230,12 +83,13 @@ def expanduser(path):
     do nothing."""
     path = os.fspath(path)
     if isinstance(path, bytes):
+        sep = b'/'
         tilde = b'~'
     else:
+        sep = '/'
         tilde = '~'
     if not path.startswith(tilde):
         return path
-    sep = _get_sep(path)
     i = path.find(sep, 1)
     if i < 0:
         i = len(path)
@@ -333,43 +187,6 @@ def expandvars(path):
             i = len(path)
             path += tail
     return path
-
-
-# Normalize a path, e.g. A//B, A/./B and A/foo/../B all become A/B.
-# It should be understood that this may change the meaning of the path
-# if it contains symbolic links!
-
-try:
-    from posix import _path_normpath as normpath
-
-except ImportError:
-    def normpath(path):
-        """Normalize path, eliminating double slashes, etc."""
-        path = os.fspath(path)
-        if isinstance(path, bytes):
-            sep = b'/'
-            dot = b'.'
-            dotdot = b'..'
-        else:
-            sep = '/'
-            dot = '.'
-            dotdot = '..'
-        if not path:
-            return dot
-        _, initial_slashes, path = splitroot(path)
-        comps = path.split(sep)
-        new_comps = []
-        for comp in comps:
-            if not comp or comp == dot:
-                continue
-            if (comp != dotdot or (not initial_slashes and not new_comps) or
-                 (new_comps and new_comps[-1] == dotdot)):
-                new_comps.append(comp)
-            elif new_comps:
-                new_comps.pop()
-        comps = new_comps
-        path = initial_slashes + sep.join(comps)
-        return path or dot
 
 
 def abspath(path):
@@ -478,81 +295,11 @@ supports_unicode_filenames = (sys.platform == 'darwin')
 
 def relpath(path, start=None):
     """Return a relative version of a path"""
-
     path = os.fspath(path)
     if not path:
         raise ValueError("no path specified")
 
-    if isinstance(path, bytes):
-        curdir = b'.'
-        sep = b'/'
-        pardir = b'..'
-    else:
-        curdir = '.'
-        sep = '/'
-        pardir = '..'
-
     if start is None:
-        start = curdir
-    else:
-        start = os.fspath(start)
+        start = b'.' if isinstance(path, bytes) else '.'
 
-    try:
-        start_tail = abspath(start).lstrip(sep)
-        path_tail = abspath(path).lstrip(sep)
-        start_list = start_tail.split(sep) if start_tail else []
-        path_list = path_tail.split(sep) if path_tail else []
-        # Work out how much of the filepath is shared by start and path.
-        i = len(commonprefix([start_list, path_list]))
-
-        rel_list = [pardir] * (len(start_list)-i) + path_list[i:]
-        if not rel_list:
-            return curdir
-        return sep.join(rel_list)
-    except (TypeError, AttributeError, BytesWarning, DeprecationWarning):
-        genericpath._check_arg_types('relpath', path, start)
-        raise
-
-
-# Return the longest common sub-path of the sequence of paths given as input.
-# The paths are not normalized before comparing them (this is the
-# responsibility of the caller). Any trailing separator is stripped from the
-# returned path.
-
-def commonpath(paths):
-    """Given a sequence of path names, returns the longest common sub-path."""
-
-    paths = tuple(map(os.fspath, paths))
-
-    if not paths:
-        raise ValueError('commonpath() arg is an empty sequence')
-
-    if isinstance(paths[0], bytes):
-        sep = b'/'
-        curdir = b'.'
-    else:
-        sep = '/'
-        curdir = '.'
-
-    try:
-        split_paths = [path.split(sep) for path in paths]
-
-        try:
-            isabs, = {p.startswith(sep) for p in paths}
-        except ValueError:
-            raise ValueError("Can't mix absolute and relative paths") from None
-
-        split_paths = [[c for c in s if c and c != curdir] for s in split_paths]
-        s1 = min(split_paths)
-        s2 = max(split_paths)
-        common = s1
-        for i, c in enumerate(s1):
-            if c != s2[i]:
-                common = s1[:i]
-                break
-
-        prefix = sep if isabs else sep[:0]
-        return prefix + sep.join(common)
-    except (TypeError, AttributeError):
-        genericpath._check_arg_types('commonpath', *paths)
-        raise
+    return posixpath.pure.relpath(abspath(path), abspath(start))
